New god: the Ieoh Jian Council
-- Gameplay Niche --

IJC is a god based on mobile melee combat, offering strong incentives
to fight groups of enemies in the open. It provides melee characters
with a strong but relatively flat power increase in the form of special
attacks triggered by moving, and offers effective positioning tools to
hybrids and ranged characters.

IJC is different from other melee focused gods in that its active
abilities are unlocked late and lacks any permanent gifts. It is instead
focused on offense and constant mobility. Its associated brand is Speed,
and to a lesser extent Flaming and Electrocution.

-- Theme --

IJC is a council of formerly mortal monks. The theme draws heavily from
Chinese imagery and traditional Chinese martial arts in particular.

-- Powers --

The god revolves around three "martial attacks" and the interactions
between them. The player is incentivized to use these special attacks
in difficult fights, and they're designed to prevent any of the three
from being overcentralizing or optimal for any particular build.

Below is a brief description of each of them, and the justification for
each particular effect.

Lunge: Triggered by moving towards an enemy. Deals extra damage against
slowed or distracted enemies.
   * It is important that constantly luring and lunging against full-speed
   monsters isn't optimal; hence, Lunge doesn't impart any status effect,
   and instead exploits monsters who are already unable to kite effectively.

Whirlwind: Triggered by moving from a tile adjacent to an enemy to another
tile adjacent to it. Can cause slow.
   * Whirlwind is the most spammable of the moves, so it is assigned a
   non-stacking, longer duration debuff so players are incentivized to
   improvise once the effect triggers. The slow lasts around six turns.

Wall Jump: Triggered by moving against a solid obstacle. It propels the
character two tiles in the opposite direction, attacking in an area of
effect. Can cause a brief distraction to onlookers.
   * Wall Jump provides the most benefit in terms of positioning and
   escape, while being the least spammable, and as such is assigned a shorter
   duration, more tactical debuff (1-3 turns).

There isn't a 1-to-1 relation between moving and attacking. Instead,
moving causes the player to trigger the appropriate martial attack a
number of times, that averages out to the same number of attacks per
aut as normal tabbing (thanks to weighted rounding).

After the three moves are unlocked, the player unlocks two active
abilities.

Serpent's Lash: Instant ability with Exhaustion and a moderate Piety
cost. Causes your next two movement actions to be instant, and boosts
your martial attacks during them. Serpent's Lash can be used as a simple
escape, but given how the geometry of the level affects Wall Jump, it's
generally more efficient to use martial attacks even when running away,
which combines well with the kit.

Heaven on Earth: Summons a storm of golden clouds. Provides a hefty
Slaying bonus and constantly triggers a powerful opaque fog, reducing
visibility and protecting against ranged monsters. The storm is very
short lived, but continuing to use martial attacks against multiple
enemies refreshes the duration and increases the slaying bonus. It
can last indefinitely if the player is able to stay in combat against
multiple enemies in the open. Heaven on Earth has explicit synergy
with Serpent's Lash, as the latter will prevent the storm from decaying
while allowing the player to build it back up quickly.